requirements.txt
asyncssh
asyncio
aiohttp
pyyaml
pylint
networkx
pyvis
colorama
jinja2